[REFACTOR]: add the needed libs
