+++ Codes/src/app/app.config.ts
@@ -1,20 +1,21 @@
-import {
-  ApplicationConfig,
-  provideBrowserGlobalErrorListeners,
-  provideZoneChangeDetection,
-} from '@angular/core';
+import { ApplicationConfig, ErrorHandler } from '@angular/core';
 import { provideRouter } from '@angular/router';
-
 import { routes } from './app.routes';
 import {
   provideClientHydration,
   withEventReplay,
 } from '@angular/platform-browser';
 
+class GlobalErrorHandler implements ErrorHandler {
+  handleError(error: unknown): void {
+    // 🔴 Log to console (or send to monitoring service)
+    console.error('Global error caught:', error);
+  }
+}
+
 export const appConfig: ApplicationConfig = {
   providers: [
-    provideBrowserGlobalErrorListeners(),
-    provideZoneChangeDetection({ eventCoalescing: true }),
+    { provide: ErrorHandler, useClass: GlobalErrorHandler },
     provideRouter(routes),
     provideClientHydration(withEventReplay()),
   ],
