+++ Codes/src/app/app.routes.ts
@@ -1,10 +1,10 @@
 // src/app/app.routes.ts
 import { Routes } from '@angular/router';
-import { LoginUser } from './pages/login-user/login-user';
-import { LoginAdmin } from './pages/login-admin/login-admin';
+import { LoginUser } from './pages/login/login-user';
+import { LoginAdmin } from './pages/login/admin/login-admin';
 
 export const routes: Routes = [
   { path: '', redirectTo: '/user-login', pathMatch: 'full' },
-  { path: 'user-login', component:  LoginUser },
+  { path: 'user-login', component: LoginUser },
   { path: 'admin-login', component: LoginAdmin },
-];   
+];
