+++ Codes/src/app/app.ts
@@ -1,13 +1,13 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { LoginAdmin } from './pages/login-admin/login-admin';
-import { LoginUser } from './pages/login-user/login-user';
+import { LoginAdmin } from './pages/login/admin/login-admin';
+import { LoginUser } from './pages/login/login-user';
 
 @Component({
   selector: 'app-root',
-  imports: [RouterOutlet,LoginAdmin,LoginUser],
+  imports: [RouterOutlet, LoginAdmin, LoginUser],
   templateUrl: './app.html',
-  styleUrl: './app.scss'
+  styleUrl: './app.scss',
 })
 export class App {
   protected title = 'movo-project';
